requirements.txt

blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pyodbc==5.2.0
Werkzeug==3.1.3
gunicorn
psycopg2-binary
